requirements.txt
attrs==22.2.0
bcrypt==4.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
defusedxml==0.7.1
exceptiongroup==1.1.1
falcon==3.1.1
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpie==3.2.1
idna==3.4
iniconfig==2.0.0
markdown-it-py==2.2.0
mdurl==0.1.2
multidict==6.0.4
packaging==21.3
pluggy==1.0.0
pydantic==1.10.7
Pygments==2.14.0
PyJWT==2.6.0
PySocks==1.7.1
pytest==7.2.2
pytest-asyncio==0.21.0
requests==2.28.2
requests-toolbelt==0.10.1
rich==13.3.2
SQLAlchemy==2.0.7
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
